+++ src/client/components/delete_btn/delete_btn.js
@@ -41,7 +41,7 @@ Template.delete_btn.onCreated( function(){
             dataContext.table.opt( 'deleteButtonTitle', pwixI18n.label( I18N, 'delete.btn_title', dataContext.item._id ), dataContext.item )
                 .then(( res ) => { self.PCK.title.set( res ); });
             // hide when disabled ?
-            this.opt( 'hideDisabled', Tabular.configure().hideDisabled )
+            dataContext.table.opt( 'hideDisabled', Tabular.configure().hideDisabled )
                 .then(( res ) => { self.PCK.hideWhenDisabled.set( res ); });
             // run only once
             comp.stop();